Refactoring package and module names for preview processor and annotations lib
