--- paparazzi-annotation-processor/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=paparazzi-annotation-processor
-POM_NAME=Paparazzi Annotation Processor
-POM_DESCRIPTION=An annotation processor for integrating Paparazzi with Composable @Previews
-POM_PACKAGING=jar
