--- paparazzi-annotation-processor/src/main/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,1 +1,0 @@
-app.cash.paparazzi.annotation.processor.PaparazziProcessorProvider
