+++ paparazzi-preview-processor/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=paparazzi-preview-processor
+POM_NAME=Paparazzi Preview Processor
+POM_DESCRIPTION=A symbol processor for integrating Paparazzi with Composable @Previews
+POM_PACKAGING=jar
