+++ paparazzi-preview-processor/src/main/java/app/cash/paparazzi/preview/processor/PreviewProcessor.kt
@@ -1,4 +1,4 @@
-package app.cash.paparazzi.annotation.processor
+package app.cash.paparazzi.preview.processor
 
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.processing.SymbolProcessor
@@ -6,11 +6,11 @@ import com.google.devtools.ksp.processing.SymbolProcessorEnvironment
 import com.google.devtools.ksp.processing.SymbolProcessorProvider
 import com.google.devtools.ksp.symbol.KSAnnotated
 
-class PaparazziProcessorProvider : SymbolProcessorProvider {
-  override fun create(environment: SymbolProcessorEnvironment) = PaparazziProcessor(environment)
+class PreviewProcessorProvider : SymbolProcessorProvider {
+  override fun create(environment: SymbolProcessorEnvironment) = PreviewProcessor(environment)
 }
 
-class PaparazziProcessor(
+class PreviewProcessor(
   private val environment: SymbolProcessorEnvironment
 ) : SymbolProcessor {
   override fun process(resolver: Resolver): List<KSAnnotated> = emptyList()
